updated cio, ciocc scripts to handle the .c -> .ll -> .mir -> .s -> .o pipeline
Now there are 1,124 errors with the mitigated MIR (probably from our stuff) about
MIR invariants we are breaking. This is not one error that is across many files, but many
different errors across a lot of our different transforms that occur in many files.
We don't have time to fix all of these, so serializing and deserializing the MIR is not
going to work.