Puts the piecewise cio compilation pipeline methodology in its own branch
This piecewise methodology is the method where do: .c -> .ll -> .mir -> .s -> .o,
analyze the object file, use the analysis results to transform (mitigate)
the saved .mir, then turn the mitigated .mir into another .o file,
and verify the .o file's safety.

This method has been pretty brittle--there are 1,142 unresolved errors
in the part of this pipeline where we transform (mitigate) the .mir
and then turn the .mir into the final .o file. These errors seem to be
from our transformations breaking many of the invariants the MIR
serializer/deserializer (printer/parser) expect to hold.

There are also bugs in the serializer/deserializer: just serializing and
then deserializing the MIR results in errors--without us even transforming
the MIR! This may actually be caused by us not running everything the normal
compilation pipeline runs over the MIR.

Saving this branch at this state in case we want to revisit this or
try to resolve the errors.

*The file cio.log contains a complete log of running this pipeline over
libsodium (1.0.18-release with out two patches applied) as well as all
of these errors.* This is saved because reproducing this requires setting
all of cio up and waiting for the first round of analysis results to finish,
so total reproducing time is ~25 minutes.